requirements.txt
absl-py==2.3.1
accelerate==1.8.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
alphashape==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.10.0
async-timeout==4.0.3
attrs==25.3.0
Authlib==1.3.1
azure-ai-formrecognizer==3.3.3
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-search-documents==11.4.0b3
azure-storage-blob==12.19.1
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.40
botocore==1.38.40
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
click-log==0.4.0
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.11
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
Crawl4AI==0.7.2
cryptography==45.0.3
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
Django==5.2.7
docling==2.39.0
docling-core==2.40.0
docling-ibm-models==3.6.0
docling-parse==4.1.0
easyocr==1.7.2
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.115.13
filelock==3.18.0
filetype==1.2.0
firecrawl==2.16.2
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-vision==3.10.2
google-genai==1.27.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-health-checking==1.73.0
grpcio-status==1.71.2
grpcio-tools==1.73.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.3
humanfriendly==10.0
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inquirerpy==0.3.4
interchange==2021.0.4
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kagglehub==0.3.12
kernels==0.9.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.45
lark==1.2.2
latex2mathml==3.78.0
lazy_loader==0.4
litellm==1.74.8
looseversion==1.3.0
lxml==5.4.0
lxml_html_clean==0.4.2
Markdown==3.8.2
markdown-it-py==3.0.0
marko==2.1.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mcp==1.13.1
mdurl==0.1.2
monotonic==1.6
mpire==2.10.2
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.4.4
multiprocess==0.70.18
mypy_extensions==1.1.0
neo4j==5.28.1
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
ocr-utils==0.1
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai==1.84.0
openai-clip==1.0.1
OpenCC==1.1.9
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pansi==2024.11.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pefile==2023.2.7
pfzy==0.3.4
pi_heif==1.0.0
pikepdf==9.9.0
pillow==11.2.1
playwright==1.54.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psutil==7.0.0
psycopg2==2.9.10
puremagic==1.30
py-cpuinfo==9.0.0
py2neo==2021.2.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==4.0.1
pyee==13.0.0
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
pylatexenc==2.10
PyMuPDF==1.26.3
pyOpenSSL==25.1.0
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.6.0
pypdfium2==4.30.1
pyperclip==1.9.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
pyxnat==1.6.3
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
rdflib==7.1.4
readability-lxml==0.8.4.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
rtree==1.4.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
semchunk==2.2.2
sentence-transformers==4.1.0
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tf-playwright-stealth==1.2.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==1.0.16
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traits==7.0.2
transformers==4.55.0
trimesh==4.7.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.155
ultralytics-thop==2.0.14
unstructured==0.18.3
unstructured-client==0.38.1
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
validators==0.34.0
wcwidth==0.2.13
weaviate==0.1.2
weaviate-client==4.16.9
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
